+++ neurokit2/ppg/__init__.py
@@ -2,14 +2,15 @@
 
 # Aliases
 from ..signal import signal_rate as ppg_rate
+from .ppg_analyze import ppg_analyze
 from .ppg_clean import ppg_clean
+from .ppg_eventrelated import ppg_eventrelated
 from .ppg_findpeaks import ppg_findpeaks
+from .ppg_intervalrelated import ppg_intervalrelated
+from .ppg_methods import ppg_methods
 from .ppg_plot import ppg_plot
 from .ppg_process import ppg_process
 from .ppg_simulate import ppg_simulate
-from .ppg_intervalrelated import ppg_intervalrelated
-from .ppg_eventrelated import ppg_eventrelated
-from .ppg_analyze import ppg_analyze
 
 __all__ = [
     "ppg_simulate",
@@ -18,6 +19,7 @@ __all__ = [
     "ppg_rate",
     "ppg_process",
     "ppg_plot",
+    "ppg_methods",
     "ppg_intervalrelated",
     "ppg_eventrelated",
     "ppg_analyze",
