--- neurokit2/report/__init__.py
@@ -1,5 +1,0 @@
-"""Submodule for NeuroKit."""
-
-from .report_create import report_create, get_default_args
-
-__all__ = ["report_create", "get_default_args"]
